+++ libc/string/memset.c
@@ -1,6 +1,6 @@
 #include <stdint.h>
 #include <sys/types.h>
-void *memset(void *s, uint_fast64_t c, size_t len) {
+void *memset(void *s, uint8_t c, size_t len) {
 	uint8_t *p = s;
 	while (len > 0) {
 		*p = c;
