+++ libc/string/strnlen_s.c
@@ -1,6 +1,6 @@
 #include <stddef.h>
-size_t strnlen_s(const char* str, size_t strsz) {
-	const char* s;
+size_t strnlen_s(const char *str, size_t strsz) {
+	const char *s;
 	for (s = str; *s && strsz--; ++s);
-	return (s - str);
+	return (size_t)(s - str);
 }
